skills/rpa-vision/requirements.txt
pyautogui>=0.9.54
mss>=9.0.1
pillow>=10.0.0
pywin32>=306
opencv-python>=4.8.0

# Windows OCR (使用 winsdk，不是 winrt-*)
winsdk>=1.0.0

# Optional OCR engines
# paddleocr>=2.7.0
# pytesseract>=0.3.10
